requirements.txt
aliyun-python-sdk-core==2.13.32
aliyun-python-sdk-kms==2.14.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
crcmod==1.7
cryptography==3.4.6
ffmpeg-python==0.2.0
Flask==1.1.2
Flask-Caching==1.10.1
future==0.18.2
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
oss2==2.14.0
pycparser==2.20
pycryptodome==3.10.1
requests==2.25.1
six==1.15.0
tqdm==4.59.0
urllib3==1.26.4
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
